requirements.txt
## Google API Python Client
google-api-python-client
## python-dateutil
python-dateutil